tedge_api: Create `Service` type
A new `Service` type was created, containing topic ids of the service
itself and the associated device. Functions which need to take an entity
topic id of a service can use this type to ensure that the given
argument is indeed a service, as well as to obtain the topic id of the
associated device.

Signed-off-by: Marcel Guzik <marcel.guzik@inetum.com>