Fix integer underflow in EntityTopicId parsing
In a typical case of a fencepost error, segments were confused with
segment separators and thus in topics with 4 slashes erroneously passed
the "topic has max of 4 segments" check. This would later cause an
underflow when generated missing topic segments, leading to a string of
size near `usize::MAX` being allocated.

Signed-off-by: Marcel Guzik <marcel.guzik@inetum.com>