fix 2311: don't publish c8y register message twice
To solve the issue, registration message conversion was changed to only
convert if the registration message affected any entities. An update is
considered to affect an entity if the entity was previously registered
and the update changes any part of entity metadata (i.e. is different
from a previous one) or adds or removes children of a device.

As part of the fix, it was necessary to change the
`EntityRegistrationMessage::payload` field. It was renamed to `other`
and made to not include `@id`, `@type`, `@parent`, etc. fields that are
already direct fields of `EntityRegistrationMessage`.

A test `does_not_duplicate_global_registration_2311` was added to ensure
the fix had intended effect. The fix was confirmed to turn the test from
red to green.

One non-intuitive part is that the registration of a brand new entity is
not considered to affect that new entity. Instead, adding or removing
any entity always affects the main device, since it's the root parent of
all other entities, so the cloud registration message is sent because
the main device is updated. In the future, after we update entity store
we might instead dynamically parse affected entities returned by the
update function to determine which cloud registration messages to send
upstream. It would then make sense to revise the decision that
registering a new entity is not considered by the entity store to affect
that entity.

Signed-off-by: Marcel Guzik <marcel.guzik@inetum.com>