tedge-log-manager: wait for upload before marking operation successful
tedge-log-manager previously marked `log_upload` operation as successful
without waiting for the actual HTTP file upload operation to finish
successfully. This lead to race conditions where after marking the
operation as successful the actual logfile may not yet be available, or
to bugs where the operation is marked as successful but the upload could
not be completed.

Now tedge-log-manager actor waits for the upload to be successful to
mark the operation as successful.