fix: use FTS for necessary operations
Makes tedge-mapper-c8y use the File Transfer Service (FTS) when
processing `log_upload` or `config_snapshot` operations.

Previously tedge-mapper-c8y implicitly assumed that it's installed on
the same device as the tedge-agent and tried to access the file-transfer
directory directly. Now it correctly downloads the files in the
operation's `tedgeUrl` parameter via HTTP.

This implementation directly uses downloader and waits until the
download is done, which means it will block if the download takes a long
time.