fix: make FTS download non-blocking
This commit changes the use of `Downloader` inside `log_upload` and
`config_snapshot` handlers to use `DownloaderActor` instead.

The advantage of this is that downloading from the FTS no longer blocks
and other such operations can be processed in the meantime.

I will argue, however, that this also had negative effects on code
complexity and maintainability. Whereas previously the download was in a
single place and one could trace the control flow easily, now the state
with maintaining the state of the download is spread very thin across
many different functions and variables.

The effect of the current architecture is that for every complex
operation where multiple steps are required, for every transition `step1
-> step2` (e.g. before the download and after the download) one has to
create:

- a piece of state where we await the response from the actor that
  performs the effect associated with the transition, e.g.
  `pending_fts_download_operations` for `DownloaderActor`
- a function `step1` which starts the transition to `step2` by sending
  a message `request`
- a function `step2` which gets the result `response`
- a function which listens to `response` from the actor, and inspects
  the response to dispatch it to correct function which performs the
  next step

I believe that Rust's async/await was created to solve, among others,
exactly this problem, by transforming every async function into a state
machine which automatically handles transistions between `.await` points
which block.