Introduce a new Scheduled state for operations
The motivation is to be able to register a custom version of a built-in
workflow. The mappers still create commands in the Init state, but the
built-in operations only react on the Schedule state. If no custom
version of the workflow has been provided, the agent moves operation
requests from the Init to the Scheduled state. If a custom version is
provided, then this user-defined workflow can add extra checks and steps
before triggering the built-in operation steps (by putting the operation
in the Schedule state).

As of now, the log and config operation are unchanged (starting on Init
and ignoring the Schedule state).

Signed-off-by: Didier Wenzek <didier.wenzek@free.fr>