Add debouncing to fs notify event stream
This commit adds debouncing of the `NotifyStream` events via the
`notify-debouncer-full` crate. The debouncing window is configured as
50ms, leading to sending of `FsEvent` messages being delayed by that
amount and duplicate events within that window being merged, which
should reduce the number of duplicate messages.

The tests had to be altered to delay deleting files, as if a file was
created and deleted within the debounce window, the `FileCreated` and
`FileRemoved` events would not fire. Other parts of the system which
rely on fs events and also create and delete files quickly, not
anticipating the delay, could also exhibit similar problems if we're not
careful. FsEvents should not be relied on too much, as they can be
unreliable[1].

A new test was added that checks if creating, modifying, deleting, or
renaming a file causes `FsEvent::Modified` to be emitted. This event is
emitted so that it can be used by consumers which need to respond to
changes of the content of the type regardless if it was modified,
deleted, moved, etc.

Also, `notify` crate itself was upgraded to the latest version.

[1]: https://docs.rs/notify/latest/notify/#known-problems