fix(#2451): Only reload when file is modified
`tedge-log-plugin` and `tedge-configuration-plugin` no longer respond to
`FsEvent::FileCreated` and only listen to `FsEvent::Modified` which is
guaranteed to be emitted if the content of the file changed, regardless
of the kind of the operation which led to this change.

As described in the comment, different editors handle editing files
differently, but only listening for modifications leads to only 1 reload
taking place with following editors i tested:
- nano
- vim
- neovim

Listening for file changes could be made more simple by using
`notify-debouncer-mini` instead, which doesn't emit precise types of
events, but only `Any` events, which only describe that there was a
change to the watched file/dir, and then it is up to the application to
scan the filesystem and perform necessary steps in reaction to that
change. For relevant plugins, that would be perfect, as the only thing
we need to know is if the file was changed, and then we always read the
file and publish the message. However, to keep our own `FsEvent` layer
working, which is also used by other components, this was not done, but
can be considered in the future, if we determine file events to be too
unreliable.