Impl `CommandBoard` to store command current states
As a first step, the command board is not connected to the main actor
of the agent. This actor will have to be updated to use the command
board as the source of truth instead of MQTT retained messages.
The latter will still be used but on very specific point in time:
- to init a new command
- to cleanup a command that has been fully processed
- to await the response of a peer
- to notify progress made by the agent on each command.

Signed-off-by: Didier Wenzek <didier.wenzek@free.fr>