Support software type for c8y advanced software management in 10.14
Since the version 10.14, c8y starts supporting "type" field for software
list and "softwareType" field for software update operation.

Change the Rust code accordingly:
1) Software list
Stop combining version and software type with the connector "::". Module
type will be declared as "type" and the version field holds its pure
version only.

2) Software update
Different behaviour if a received operation contains "softwareType"
field or not. If yes, softwareType will be used as its module type and
version will be used as it is. If no, mapper tries to get the module
type from the received version field as we initially implemented.

Signed-off-by: Rina Fujino <rina.fujino.23@gmail.com>