Let tedge_operation_converter use an unbounded receiver
The tedge_operation_converter is sending messages to the
software_manager actor, while receiving messages from this same actor.
This leads to a deadlock if one their message boxes start to be full.
This is notably the case on startup if there are too many pending
operations awaiting over MQTT to be processed. With the tedge_operation_converter
using an unbounded receiver, all these messages can be consumed without
blocking status messages received from the software_manager actor.

Signed-off-by: Didier Wenzek <didier.wenzek@free.fr>