Split the FirmwareManagerMessageBox into a Receiver and a Worker
The idea is to have a worker which is Clone with methods that can be
spawned in background tasks. The actor itself keeps only stateful things that
cannot be shared (the message receiver and the cache of pending
operations).

This is an intermediate step. All the methods are for now still attached
to the actor. The next step will be to move out from the actor to the
worker all the methods that doesn't touch the actor state.

Signed-off-by: Didier Wenzek <didier.wenzek@free.fr>