Remove impl ServiceProvider for ServerActorBuilder
Connecting a server actor to its clients has been the main motivation
for the ServiceProvider trait which where used to let the client and the
server exchange senders, one given by the server to the client requests and
another given by the client to server responses.

With the new mechanism of RequestEnvelope, this exchange is no more
useful. The actor server is a simple MessageSink accepting requests
wrapped in an envelope with a reply-to address, i.e. a sender.
Clients only needs a sender to post their request.

The next step will be to fully deprecate the ServiceProvider trait,
in favor of the combination of two traits MessageSource + MessageSink.
Indeed, the complexity of the ServiceProvider was only used by server
actors and not the actors, such as the MQTTActor, for which the messages
received and sent don't have a request-response relationship.

Signed-off-by: Didier Wenzek <didier.wenzek@free.fr>