Deprecate MessageSource::register_peer
The purpose of the `register_peer` and `connect_sink` methods was
exactly the same: to connect a source to a sink. The only difference was
on their arguments, the sink being respectively passed as a sender or as
a sink. Simply adding an `impl<M> MessageSink<M> for DynSender<M>`,
makes the two methods fully equivalent.

Signed-off-by: Didier Wenzek <didier.wenzek@free.fr>