Deprecate ServiceProviderExt and new_client_box test helpers
The new_client_box method has not been fully deprecated but moved to the
Service trait. This makes this trait a bit more contrieved (with 3 ways
to connect clients) and will have to be fixed.
However, I prefer this temporary situation because moving
the method in a trait Ext was only hidding the issue.
I will address that in a following step.

Signed-off-by: Didier Wenzek <didier.wenzek@free.fr>