Add active heartbeat timer map to converter
In runtime, if user updates entity metadata message to change the health
endpoint, there can be this situation: one timer exists for the old
health endpoint, and the other timer exists for the new endpoint. To
avoid the old timer fires and gets triggered again, converter needs to
know what is the map, which entity topic ID is interested in which
health endpoint now.

Signed-off-by: Rina Fujino <rina.fujino.23@gmail.com>