create an OperationHandler struct
Creates a new `OperationHandler` struct with methods that handle
operations. This is done to both pull stuff out of
`CumulocityConverter`, which is too large and does too many things, and
to make it possible to handle operations concurrently.

The current design of the `CumulocityConverter` is that all the incoming
MQTT messages go into the converter's `convert()` method, which converts
in both directions between messages from the cloud and local MQTT
broker, and also needs to return a converted message without blocking.

In contrast, when handling operations, it can take some time before an
MQTT response is ready to be sent. As such, it makes sense to spawn a
task when starting the operation, and have the converter return no
messages, so other messages can be converted.

Signed-off-by: Marcel Guzik <marcel.guzik@inetum.com>