Support Cumulocity Device Availability
If c8y.availability.enable is set to true, tedge-mapper-c8y publishes
a SmartREST '117' message to set c8y_RequiredAvailability fragment for
both main and child devices.

If c8y.availability.enable is set to true AND c8y.availability.period is
set to value > 0, tedge-mapper-c8y sends an empty inventory update
message periodically.

For the main device, the heartbeat timer starts during initialization.
When the timer fires, the health status of tedge-agent is checked. If
it is up, the heartbeat message will be sent.

For child devices, the heartbeat timer starts when the entity is
registered to the entity store. By default, like the main device case,
the heartbeat message will be sent when tedge-agent for the child device
is up. However, if `at_health` (EntityTopicId) is set to its metadata,
tedge-mapper-c8y looks up the status of the leadService.

Signed-off-by: Rina Fujino <rina.fujino.23@gmail.com>