Add span information for `to_string` in `tedge_config_macros`
Previously, errors were spanned to the macro invocation, so it was
impossible to determine from the error alone which field was
problematic. This change ensures the error is scoped to the type of the
field in question, like we do for `FromStr` already, so it is easy to
identify the issue.

Signed-off-by: James Rhodes <jarhodes314@gmail.com>