Do not defer execution of pending commands on restart
Before, the pending commands were queued for execution on restart
of the agent. Similarly, resuming an invoking operation on sub-operation
termination was queued for execution.

This queuing approach was buggy because two events were queued and both
executed in the a context where the sub-command was known as terminated:
- termination of the sub-command
- continuation of the invoking command

Now, when a command is resumed (on restart or when a sub-operation
terminates), the command is processed immediately.

This leads to the following thread of execution:

--[ issue-2896 @ init
--[ issue-2896 @ executing
--[ issue-2896 @ await_restart
--[ issue-2896 > restart @ init
--[ issue-2896 > restart @ scheduled
--[ issue-2896 > restart @ executing  # trigger the reboot
--[ issue-2896 @ await_restart        # resumed on restart
--[ issue-2896 > restart @ executing  # resumed on restart
--[ issue-2896 > restart @ successful # detect reboot success
--[ issue-2896 @ await_restart        # proceed on reboot success
--[ issue-2896 @ restarted
--[ issue-2896 @ successful

Signed-off-by: Didier Wenzek <didier.wenzek@free.fr>