Avoid overlapping subscriptions to MQTT topics
When a message matches several subscriptions of a connection,
the very same message might be delivered several time to the connection.
This is notably the case with NanoMQ (while Mosquitto delivers the
message just once).

The proposed overlap-detection mechanism doesn't remove all the
overlaps, but only those for which holds a sub-pattern/super-pattern
relationship (e.g. te/+/+/+/+/m/+ includes te/device/main///m/+).

Signed-off-by: Didier Wenzek <didier.wenzek@free.fr>