fix configuration set operation Robot test
Test `Set Configuration when file exists` is changed to be consistent
with its comment that suggests that permissions for existing files
should not change. The test now actually asserts that permissions of
`/etc/config1.json` and `/etc/binary-config1.tar.gz` stay unchanged.

Previously, the test seemed to actually assert if the permission did
change, despite the comment claiming otherwise. Changing of permission
was not desirable, and along with resetting the owner of the file, was
part of a bug that is now fixed, so either the test must've been wrong
to begin with, or tedge-write behaviour is specified incorrectly.

`/etc/config1.json` was created with permissions 664, but the test
verified that they were 644 after the operation, which isn't consistent
with what's in `tedge-configuration-plugin.toml`, but is the default
file permissions with default umask 002, so a new file being created
without any changes to default permissions fulfilled the requirements of
the test.

Signed-off-by: Marcel Guzik <marcel.guzik@inetum.com>