move tests of leftover operations into their own modules
Tests that verify the execution of the operations (i.e. where we
generate an operation with a given id and verify its outputs) were moved
into their own modules.

As handling the operation usually consists of 2 things:
1. convert cumulocity smartrest message to local MQTT operation message
   and back
2. handle local MQTT operation message

These steps are performed by different components, so the unit tests are
also separated.

Signed-off-by: Marcel Guzik <marcel.guzik@inetum.com>