fix: stop OperationHandler spawning extra tasks
When receiving multiple MQTT messages on a command topic,
`OperationHandler::handle` method spawned extra tokio tasks, because
`.unwrap_or` and not `.unwrap_or_else` was used when checking if a task
for that topic already existed. Fortunately because returned object, and
along it the sender to the task, was immediately dropped, the task
exited and so we weren't leaking anything.

Signed-off-by: Marcel Guzik <marcel.guzik@inetum.com>