ci: update codecov range
In 80..100 the 100 is when the coverage should be marked as green.
Because we want to have nice things and would rather prefer codecov not
to complain in PR comments about files where coverage is already pretty
high, we'll be decreasing this threshold.

Also changing precision to 2 because it's the default and why not.

Also change comment layout to the "most up to date one" according to
codecov.

Signed-off-by: Marcel Guzik <marcel.guzik@inetum.com>