impl TryFrom<(TomlExitHandlers, DefaultHandlers)> for Handlers
With the previous impl TryFrom<TomlExitHandlers> for Handlers many
definition consistency and completness checks were not feasible,
the default values being injected only once the struct fully built.

This commit is a preparation step.
The next step is to deprecate the `with_default` method,
moving its effects into to the TryFrom implementations.

Signed-off-by: Didier Wenzek <didier.wenzek@free.fr>