fix issues in atomically_write_file_* functions
- no longer fail when `.tmp` file already exists, new tempfile are
  created with 6 random bytes and names are regenerated if such name
  already exists
- ensure async tempfile is cleaned up by wrapping it in a
  `tempfile::NamedTempFile` that provides a drop implementation

`tempfile` was also bumped because `tempfile::Builder::permissions` was
not available on the current version.

Running `configuration_operation.robot` suite, it was flaking for me,
and I have found the following error that caused `tedge-agent` to never
start:

```
ERROR sm-agent: tedge_actors::runtime: Aborted due to Writing the
content to the file "/etc/tedge/.agent/software-current-operation"
failed: "could not create the temporary file \"/etc/tedge/.agent/software-current-operation.tmp\"".
source=Os { code: 17, kind: AlreadyExists, message: "File exists" }`
```

If a `software-current-operation` already had a temporary file, the
atomic copy would fail. As it used async version of the atomic write
function, which was trying to delete tempfiles manually on errors, it
was nevertheless possible for this function to be terminated while
running if runtime finished, leading to tempfile being not cleaned up on
Drop.

Signed-off-by: Marcel Guzik <marcel.guzik@inetum.com>