fix: Remove hardcoded tedge-write path
API for spawning a tedge-write process no longer uses a hardcoded
`/usr/bin/tedge-write` path.

The path was hardcoded because of the wrong assumption of how `sudo`
works: when provided with a command without the full path, e.g. `sudo
tedge-write /etc/config1`, `sudo` will internally look for `tedge-write`
in $PATH and then compare path with its sudoers entry, so it's not
necessary to use full path for binaries used with `sudo`.

Signed-off-by: Marcel Guzik <marcel.guzik@inetum.com>