Stop chowning in tedge-agent
Stop setting file ownership in tedge-agent, which runs as `tedge` user.
Also added a missing test for a case when a file is owned by root.

Setting file ownership is a privileged operation, which should be done
by tedge-write. It's also unnecessary to change permissions on a
temporary file which is atomically moved into destination by
config-manager anyway.

Signed-off-by: Marcel Guzik <marcel.guzik@inetum.com>