Create SmartrestPayload type
Introduce a SmartrestPayload type, meant to encode assumptions about
payload of SmartREST messages, i.e. that:

- it is a "CSV-like" format
- it has a size limit

This type should be used in places where we previously just used a plain
String or constructed smartrest messages directly with format!, to
ensure that above invariants always hold, and as the diff shows, in some
places these invariants were being ignored. Most of them are corner
cases, i.e. where it is technically possible for a device id or template
id to increase the payload size over the limit, which should not
normally happen.

One more advantage is also that we can more easily see where smartrest
payloads are being constructed or used by looking up references to
SmartrestPayload type in an IDE or rust-analyzer.

Signed-off-by: Marcel Guzik <marcel.guzik@inetum.com>