Compare health check responses with 1s precision
If a unix timestamp returned by the service is an integer (i.e. it does
not have a fractional, sub-second part), then the comparison if the
response is newer than health check request could fail.

Thus the comparison was changed to compare with a 1s precision. The case
when timestamp is an integer works and it shouldn't return false
positives because we don't send health check requests more often than
1s.

Signed-off-by: Marcel Guzik <marcel.guzik@inetum.com>