Add --log-level flag
Add a --log-level flag that sets the verbosity of reported logs.

In contrast to --debug, which enables debug logs in addition to the
default error, warn, and info logs, --log-level sets the maximum
verbosity, i.e. when set to error print only errors, if set to warn
print warnings and errors, etc. and when set to trace, print all the
logs.

If --log-level is selected, this value is used and takes precedence over
--debug. RUST_LOG env variable still overrides the flag.

Signed-off-by: Marcel Guzik <marcel.guzik@inetum.com>