Combine C8yEndPoint with ProxyUrlGenerator
This is an intermediate step.
C8yEndPoint has been enriched with a ProxyUrlGenerator
and local_proxy_url method.
But the clients has not been changed beyond the miminum to get the
C8yEndPoint. They are still using a combination of C8yEndPoint and ProxyUrlGenerator
to generate the urls to the local auth proxy. The next step will be to
make the ProxyUrlGenerator private and enforce the use of
C8yEndPoint::local_proxy_url()

Signed-off-by: Didier Wenzek <didier.wenzek@free.fr>