Remove one-level of actors
The C8YHttpProxyActor was adding a level of indirection with little
value. Message generated by C8YHttpProxy was pushed to C8YHttpProxyActor
and then to an HttpActor.

This commit is an intermediate step. The C8YHttpProxyActor is no more an
actor but has been removed. The next step will be to merge C8YHttpProxy
and C8YHttpProxyActor.

Signed-off-by: Didier Wenzek <didier.wenzek@free.fr>