Simplify the retry logic of the c8y http proxy
The retry logic was unnecessary complicated (with nested loops, caches
and cache invalidation. As a first step these loops are replaced by
simple sequences: get the c8y internal id for the device, then use this
id to proceed (i.e. to post a software list or to create an event).

Signed-off-by: Didier Wenzek <didier.wenzek@free.fr>