Add AvailabilityActor channel deadlock test
This test is meant to verify the fix to #3279.

Another test like this will probably have to be added to cover the
original behaviour that motivated connecting AvailabilityActor to the
C8yMapperActor instead of MqttActor directly, i.e. AvailabilityActor
sending a Smartrest 117 message before C8yMapperActor sends 101 message
that registers a child device (which is invalid and should not happen).

Signed-off-by: Marcel Guzik <marcel.guzik@inetum.com>