Persist workflow definition of pending commands on start
When the agent restarts and loads pending operations,
these operations might not have been assign a workflow version.
This is notably the case after a self update from a previous version of
the agent that was oblivious of workflow versions.

In such a case, the agent assumes the current user-defined version of
the workflow is the version to be used for the pending command.
The workflow definition is persisted in `workflows-in-use` directory
and assigned to the command.

Signed-off-by: Didier Wenzek <didier.wenzek@free.fr>