Remove hard errors on BridgeConfig::validate()
Adding cryptoki configuration to the method is non-trivial and the
method IMO doesn't provide much value anyway, so I decided to replace
hard errors with warnings.

BridgeConfig struct is based on Mosquitto configuration file, so adding
fields related to cryptoki would make it a mess. If the files do
not exist, then the bridge will report it when it tries to start.

Signed-off-by: Marcel Guzik <marcel.guzik@cumulocity.com>