Remove from C8yMapperActor the publish_messages receiver
The motivation, for this receiver of wrapped MqttMessages, was that:
"other actors may want to send MqttMessages but we want to ensure that the entity referred to in the message
topic exists; that's why we route them to the mapper rather than to MqttActor directly."

However, the C8yMapperActor was sending these messages unchanged.
And futhermore, this was requiring a specific mechanism to avoid
deadlocks. Better to remove all that, and let the other actors
(actually only the AvailabilityActor) send their messages directly over
MQTT.

Signed-off-by: Didier Wenzek <didier.wenzek@free.fr>