Add TEDGE_TRACE_DIR envvar for outputting a trace file
`TEDGE_TRACE_DIR` environment variable is added to enable trace file
output in addition to already present `--trace-json`. It also allows the
user to specify the directory where the trace file will be written.

I decided on the environment variable because it may be tricky to add
the flag to all of the invocation of thin-edge services, particularly if
we want to trace all the services and when they are started indirectly,
like by systemd.

Signed-off-by: Marcel Guzik <marcel.guzik@cumulocity.com>