Add compatibility with systemd socket activation
Instead of always binding the socket, check if any sockets are passed by
systemd using sd_listen_fds. If there are, use the first one. If not,
fallback to creating a socket manually.

The service is started by systemd's socket, but otherwise operates
normally, i.e. runs the accept() loop handling multiple requests. This
corresponds to `Accept=no` systemd socket configuration. Support for
`Accept=yes` will be added in the next commit.

A new dependency sd-listen-fds was added, that only contains the
required function, but it can be replaced with libsystemd in the future.

Signed-off-by: Marcel Guzik <marcel.guzik@cumulocity.com>