tedge-p11-server: Select token using serial number
Allow the user to select a token via a serial number. If a serial number
is provided, we attempt to select a token with that number and fail if
such token doesn't exist.

As of now this is implemented for tedge-p11-server, so works with
`socket` cryptoki mode, but doesn't work yet with `module` mode.

Signed-off-by: Marcel Guzik <marcel.guzik@cumulocity.com>