refactor: mqtt auth config use Cloud trait
This is a small refactor that reduces the diff in a next commit that
contains the functional change.

Instead of using a particular C8y config reader, use the Cloud trait to
read common auth related properties like paths to certs. This reduces
duplication when we add Aws and Azure readers in the next commit.

Signed-off-by: Marcel Guzik <marcel.guzik@cumulocity.com>