Add feature to disable AWS integration
To remove the AWS dependencies, the dependencies must not be used by the
CLI and the bridge code anymore.

Use the aws feature to use conditional compilation for the AWS related
code and add it only if AWS support is enabled.

Enable the aws feature by default to retain the previous behavior.

Signed-off-by: Michael Tretter <m.tretter@pengutronix.de>