Add feature to disable Azure integration
To remove the Azure dependencies, the dependencies must not be used by
the CLI and the bridge code anymore.

Use the azure feature to use conditional compilation for the Azure
related code and add it only if Azure support is enabled.

Enable the azure feature by default to retain the previous behavior.

Signed-off-by: Michael Tretter <m.tretter@pengutronix.de>