Refactor if let with match for cloud backend
The "if let" for detecting if the C8y cloud is used prevents conditional
compilation of the C8y cloud support using attributes.

Replace the if let statements with matches which explicitly handle the
other supported clouds as well. This allows to add an attribute to each
match arm.

Another option would be to use the cfg! macro, but explicitly listing
the supported backend leads to a more consistent style across the code
in this module.

Signed-off-by: Michael Tretter <m.tretter@pengutronix.de>