test: Connect to C8y using all TLS1.3 ECDSA curves
This test is part of adding support to ECDSA384 and ECDSA521 keys.
Currently we only support ECDSA256, because ecdsa_secp256r1_sha256
scheme is offered by C8y, as well as the others. But this can change in
the future, so we should support all of the EC schemes, with the bonus
that users that wish to use bigger keys can do so.

Signed-off-by: Marcel Guzik <marcel.guzik@cumulocity.com>